requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlmodel==0.0.16
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
pydantic==2.6.3
python-multipart==0.0.9
joblib==1.3.2
matplotlib==3.8.3
seaborn==0.13.2
python-dotenv==1.0.1
PyYAML==6.0.1
